requirements.txt
Flask 
Flask-MySQLdb 
Flask-Bcrypt 
Flask-Limiter
python-dotenv